Big SCSS refactor (but still WIP).
This is a big refactor of the scss-file structure,
partials and classes (renaming still in progress).

It also refreshes the layout a bit, especially buttons,
footer and the hire button.

It also provides a temporary fix for the mobile navigation which
was not shown anymore in iOS8 due to some weird :target / :active
bug in the new Safari 8.0/8.1.

Also attached a few screenshots from recent work to be used on
the website later on in the work section.